:bug: Baseurl 오류로 인한 canonical 에러 수정
