+++ apps/blog/src/shared/routes/index.ts
@@ -1,0 +1,4 @@
+export const ROUTES = {
+  root: () => "/",
+  postDetail: (postsPath: string) => `/${postsPath}`,
+};
