+++ biome.json
@@ -56,7 +56,8 @@
       "**/coverage/*",
       "**/dist/*",
       "**/.next/*",
-      "**/node_modules/*"
+      "**/node_modules/*",
+      "*storybook.log"
     ]
   }
 }
